feat: refactor dialogs
